+++ src/low/v7400/array_attribute.rs
@@ -31,6 +31,7 @@ impl ArrayAttributeEncoding {
 
     /// Returns the raw value.
     #[cfg(feature = "writer")]
+    #[cfg_attr(feature = "docsrs", doc(cfg(feature = "writer")))]
     pub(crate) fn to_u32(self) -> u32 {
         match self {
             ArrayAttributeEncoding::Direct => 0,
@@ -55,7 +56,7 @@ impl FromReader for ArrayAttributeEncoding {
     fn from_reader(reader: &mut impl io::Read) -> Result<Self, ParserError> {
         let raw_encoding = u32::from_reader(reader)?;
         let encoding = ArrayAttributeEncoding::from_u32(raw_encoding)
-            .ok_or_else(|| DataError::InvalidArrayAttributeEncoding(raw_encoding))?;
+            .ok_or(DataError::InvalidArrayAttributeEncoding(raw_encoding))?;
         Ok(encoding)
     }
 }