+++ src/writer/v7400/binary.rs
@@ -89,13 +89,6 @@
 //!
 //! # Ok::<_, fbxcel::writer::v7400::binary::Error>(())
 //! ```
-//!
-//! [`AttributesWriter`]: struct.AttributesWriter.html
-//! [`Writer::close_node`]: struct.Writer.html#method.close_node
-//! [`Writer::finalize`]: struct.Writer.html#method.finalize
-//! [`Writer::finalize_and_flush`]: struct.Writer.html#method.finalize_and_flush
-//! [`Writer::new`]: struct.Writer.html#method.new
-//! [`Writer::new_node`]: struct.Writer.html#method.new_node
 
 use std::{
     convert::TryFrom,
@@ -120,7 +113,7 @@ mod footer;
 
 /// Binary writer.
 ///
-/// See [module documentation](index.html) for usage.
+/// See [module documentation][`self`] for usage.
 #[derive(Debug, Clone)]
 pub struct Writer<W: Write> {
     /// Writer destination.
@@ -218,7 +211,7 @@ impl<W: Write + Seek> Writer<W> {
             return Ok(());
         }
 
-        let current_pos = self.sink.seek(SeekFrom::Current(0))?;
+        let current_pos = self.sink.stream_position()?;
         current_node.header.bytelen_attributes = current_pos - current_node.body_pos;
         current_node.is_attrs_finalized = true;
 
@@ -244,7 +237,7 @@ impl<W: Write + Seek> Writer<W> {
         let bytelen_name =
             u8::try_from(name.len()).map_err(|_| Error::NodeNameTooLong(name.len()))?;
 
-        let header_pos = self.sink.seek(SeekFrom::Current(0))?;
+        let header_pos = self.sink.stream_position()?;
 
         let header = NodeHeader {
             end_offset: 0,
@@ -259,7 +252,7 @@ impl<W: Write + Seek> Writer<W> {
         // Write node name.
         self.sink.write_all(name.as_ref())?;
 
-        let body_pos = self.sink.seek(SeekFrom::Current(0))?;
+        let body_pos = self.sink.stream_position()?;
 
         self.open_nodes.push(OpenNode {
             header_pos,
@@ -288,7 +281,7 @@ impl<W: Write + Seek> Writer<W> {
         }
 
         // Update node header.
-        let node_end_pos = self.sink.seek(SeekFrom::Current(0))?;
+        let node_end_pos = self.sink.stream_position()?;
         self.sink.seek(SeekFrom::Start(current_node.header_pos))?;
         current_node.header.end_offset = node_end_pos;
         assert_eq!(
@@ -304,6 +297,7 @@ impl<W: Write + Seek> Writer<W> {
 
     /// Writes the given tree.
     #[cfg(feature = "tree")]
+    #[cfg_attr(feature = "docsrs", doc(cfg(feature = "tree")))]
     pub fn write_tree(&mut self, tree: &crate::tree::v7400::Tree) -> Result<()> {
         use crate::low::v7400::AttributeValue;
 
@@ -370,9 +364,7 @@ impl<W: Write + Seek> Writer<W> {
 
     /// Finalizes the FBX binary and returns the inner sink.
     ///
-    /// You may want to use [`finalize_and_flush()`].
-    ///
-    /// [`finalize_and_flush()`]: #method.finalize_and_flush
+    /// You may want to use [`finalize_and_flush()`][`Self::finalize_and_flush()`].
     pub fn finalize(mut self, footer: &FbxFooter<'_>) -> Result<W> {
         self.finalize_impl(footer)?;
 
@@ -401,7 +393,7 @@ impl<W: Write + Seek> Writer<W> {
         {
             let len = match footer.padding_len {
                 FbxFooterPaddingLength::Default => {
-                    let current = self.sink.seek(SeekFrom::Current(0))?;
+                    let current = self.sink.stream_position()?;
                     current.wrapping_neg() & 0x0f
                 }
                 FbxFooterPaddingLength::Forced(len) => u64::from(len),